feat: 서버 통신 로그 볼 수 있는거 추가
