+++ app/build.gradle
@@ -59,4 +59,9 @@ dependencies {
     // Retrofit
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
+
+    // OkHttp
+    implementation group: 'com.squareup.okhttp3', name: 'okhttp', version: '4.5.0'
+    // Use for HttpLoggingInterceptor
+    implementation group: 'com.squareup.okhttp3', name: 'logging-interceptor', version: '4.5.0'
 }
