+++ app/src/main/java/com/snowflowerthon/snowman/data/ApiService.kt
@@ -4,13 +4,17 @@ import com.snowflowerthon.snowman.data.dto.BaseResponseDto
 import com.snowflowerthon.snowman.data.dto.request.VoteRequsetDto
 import retrofit2.Call
 import retrofit2.http.Body
+import retrofit2.http.Header
 import retrofit2.http.POST
+import retrofit2.http.Path
 import retrofit2.http.Query
 
 
 interface ApiService {
-    @POST("/api/v1/weathers/{weatherId}/poll")
-    fun voteClothes(@Query("weatherId") weatherId: Long,
+    @POST("/api/v1/weathers/{regionId}/poll")
+    fun voteClothes(@Header("Authorization") token: String,
+                    @Path("regionId") regionId: Long,
+                    @Query("userId") userId: Long,
                     @Body request: VoteRequsetDto)
             : Call<BaseResponseDto<String?>>
