--- app/src/main/java/com/snowflowerthon/snowman/data/Retrofit.kt
@@ -1,23 +1,0 @@
-package com.snowflowerthon.snowman.data
-
-import com.snowflowerthon.snowman.BuildConfig.BASE_URL
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class RetrofitClient {
-    // 객체를 하나만 생성하는 싱글톤 패턴을 적용합니다.
-    companion object {
-        // API 서버의 주소가 BASE_URL이 됩니다.
-        private var INSTANCE: Retrofit? = null
-
-        fun getInstance(): Retrofit {
-            if(INSTANCE == null) {  // null인 경우에만 생성
-                INSTANCE = Retrofit.Builder()
-                    .baseUrl(BASE_URL)  // API 베이스 URL 설정
-                    .addConverterFactory(GsonConverterFactory.create()) // 1)
-                    .build()
-            }
-            return INSTANCE!!
-        }
-    }
-}
